requirements.txt
requests
pytz
pyserial
firebase-admin
configargparse

